+++ src/by/bsu/travelagency/entity/User.java
@@ -1,8 +1,5 @@
 package by.bsu.travelagency.entity;
 
-/**
- * Created by Михаил on 2/24/2016.
- */
 public class User extends Entity {
     
     /** The id. */
@@ -30,7 +27,7 @@ public class User extends Entity {
     private double discount;
     
     /** The money. */
-    private int money;
+    private double money;
 
     /**
      * Instantiates a new user.
@@ -187,7 +184,7 @@ public class User extends Entity {
      *
      * @return the money
      */
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
@@ -196,7 +193,7 @@ public class User extends Entity {
      *
      * @param money the new money
      */
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 }
